Add build targets (#9)
* Use parallel-webpack

* Add build targets

* Add webpack dev config

* Build with parallel-webpack

* Expose util functions for usage in non React environment

* Update README
