+++ src/utils.js
@@ -1,0 +1,2 @@
+export { createElement } from 'react';
+export { render } from 'react-dom';
